requirements.txt
pandas
numpy
telethon
seaborn
matplotlib
python-dotenv